% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dance.R
\name{dance_recital}
\alias{dance_recital}
\title{Create a matahari-esque data frame from a file.}
\usage{
dance_recital(code, evaluate = TRUE)
}
\arguments{
\item{code}{A string or the path to a file containing R code.}

\item{evaluate}{Logical, indicating whether to evaluate the code, default is \code{TRUE}}
}
\description{
Create a matahari-esque data frame from a file.
}
\examples{
code_file <- system.file("test", "sample_code.R", package = "matahari")
dance_recital(code_file)
}
