% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dance.R
\name{dance_report}
\alias{dance_report}
\title{Copy the log to your clipboard.}
\usage{
dance_report(...)
}
\arguments{
\item{...}{Developer options.}
}
\description{
Copy the log to your clipboard.
}
\examples{
\dontrun{

# Start recording
dance_start(value = TRUE)

# Execute some expressions
"Hello!"
4 + 4

# Stop recording
dance_stop()

# Assign a base64 encoded tibble of your log to a variable
report_code <- dance_report()
substr(report_code, 1, 10)

## "WAoAAAADAA"

nchar(report_code)

## 397432
}
}
