% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dance.R
\name{dance_save}
\alias{dance_save}
\title{Save the log as an rds file.}
\usage{
dance_save(path)
}
\arguments{
\item{path}{The path to the rds file.}
}
\description{
Save the log as an rds file.
}
\examples{
\dontrun{

# Start recording
dance_start(value = TRUE)

# Execute some expressions
"Hello!"
4 + 4

# Stop recording
dance_stop()

# Save your log locally
dance_save("session.rds")
}
}
