% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dance.R
\name{dance_start}
\alias{dance_start}
\title{Start Spying}
\usage{
dance_start(expr = TRUE, value = FALSE, path = FALSE, contents = FALSE,
  selection = FALSE)
}
\arguments{
\item{expr}{The R expressions typed into the R console will be logged unless
this is set to \code{FALSE}.}

\item{value}{Values that are computed on the R console will be logged unless
this is set to \code{FALSE}.}

\item{path}{The path to the file in focus on the RStudio editor will be logged
unless this is set to \code{FALSE}.}

\item{contents}{The file contents of the RStudio editor tab in focus will be
logged unless this is set to \code{FALSE}.}

\item{selection}{The text that is highlighted in the RStudio editor tab in
focus will be logged unless this is set to \code{FALSE}.}
}
\description{
Start logging your R console activity. If you're using RStudio the contents
of your current editor tab will also be tracked. All logging takes place when
R code is executed in the R console.
}
