% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dance.R
\name{dance_stop}
\alias{dance_stop}
\title{Stop Spying}
\usage{
dance_stop()
}
\value{
\code{TRUE} if logging was taking place, otherwise \code{FALSE} (invisibly).
}
\description{
Pause the current logging session.
}
\examples{
\dontrun{
# Begin recording your R session.
dance_start()

# Each expression adds a row to the tibble that tracks the execution of
# your R code.
x <- 7
x * 4

# This stops recording your R session.
dance_stop()
}
}
