% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dance.R
\name{dance_recital}
\alias{dance_recital}
\title{Create a matahari-esque data frame from a file.}
\usage{
dance_recital(code)
}
\arguments{
\item{code}{A string or the path to a file containing R code.}
}
\description{
Create a matahari-esque data frame from a file.
}
