% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dance.R
\name{dance_remove}
\alias{dance_remove}
\title{Remove your logging history.}
\usage{
dance_remove()
}
\value{
Either \code{TRUE} if the log was removed or \code{FALSE} if the log
does not exist (invisibly).
}
\description{
Remove your logging history.
}
\examples{
\dontrun{

dance_start()
"Hello!"
4 + 4
dance_stop()
dance_tbl()
dance_remove()
dance_tbl()
}
}
