% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dance.R
\name{dance_report}
\alias{dance_report}
\title{Copy the log to your clipboard.}
\usage{
dance_report(...)
}
\arguments{
\item{...}{Developer options.}
}
\description{
Copy the log to your clipboard.
}
\examples{
\dontrun{

dance_start()
"Hello!"
4 + 4
dance_stop()
dance_report()
}
}
