% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dance.R
\name{dance_save}
\alias{dance_save}
\title{Save the log as an rds file.}
\usage{
dance_save(path)
}
\arguments{
\item{path}{The path to the rds file.}
}
\description{
Save the log as an rds file.
}
\examples{
\dontrun{

dance_start()
"Hello!"
4 + 4
dance_stop()
dance_save("session.rds")
}
}
