% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dance.R
\name{dance_stop}
\alias{dance_stop}
\title{Stop Spying}
\usage{
dance_stop()
}
\value{
\code{TRUE} if logging was taking place, otherwise \code{FALSE} (invisibly).
}
\description{
Pause the current logging session.
}
\examples{
\dontrun{

dance_start()
"Hello!"
4 + 4
x <- 7
x^2
rm(x)
x
dance_stop()
}
}
