% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dance.R
\name{dance_tbl}
\alias{dance_tbl}
\title{Get the log as a tibble}
\usage{
dance_tbl()
}
\value{
Either a \code{\link[tibble:tibble]{tibble::tibble()}} containing your logged history or \code{NULL}.
if there is no log.
}
\description{
Get the log as a tibble
}
\examples{
\dontrun{

dance_start()
"Hello!"
4 + 4
dance_stop()
dance_tbl()
}
}
