% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dance.R
\name{dance_tbl}
\alias{dance_tbl}
\title{Get the log as a data frame}
\usage{
dance_tbl()
}
\value{
Either a data frame containing your logged history or \code{NULL}
if there is no log.
}
\description{
Get the log as a data frame
}
